% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data Set}
\format{A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{Distance from fire station in miles}
  \item{DAMAGE}{x 1000 dollars in damage}
  ...
}}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences-Sixth-Edition/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A dataset with 2 variables both of which are quantitative.
Interest is in the amount of Damage in dollars and distance from the fire station in miles
}
\keyword{datasets}
