% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{- number of iterations}

\item{n}{- number}

\item{p}{- probability}
}
\value{
succ.tab/iter
}
\description{
Creates a table of successes in binomial distribution over iter number of trials.
Fills a matrix and creates a labeled barplot for given input.
}
\examples{
mybin(iter=100,n=10, p=0.5)
}
