% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{My Central Limit Theorem Binomial Function}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{- size of sample}

\item{iter}{- number of iterations}

\item{p}{- The probability of success in a single binomial trial}

\item{...}{}
}
\value{
a histogram of sample mean theoretical normal curve line added in Red
}
\description{
mycltb takes a random sample from a binomial distribution.
CLT will work with discrete OR continuous distributions
}
\examples{
mycltb(n=3, iter=10000, p=0.5)
mycltb(n=9, iter=10000)

}
