% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{a}{}
}
\value{
probability - area under the curve
}
\description{
display the curve, shaded area between the curve and x axis
from -∞ to x=a, and calculate the area (probability, P(X<=a))
which is released to the command-line in a list.
}
\examples{
myncurve(2,4,5)
}
