% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin - Binomial Simulation}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{iter}{- number of iterations}

\item{n}{- sample size (number of trials)}

\item{p}{- probability of a success}
}
\value{
plot of the simulation with proportions and a table of proportions
}
\description{
Uses sample function to produce a binomial simulation over iter number of trials.
Fills a matrix and creates a labeled barplot for given input.
}
\examples{
mybin(iter=100,n=10, p=0.5)
}
